TEMP: CHROMIUM: ASoC: samsung: reparent mout_i2s for suspend/resume
On reset, fin_pll is the parent of mout_audss and mout_audss is the
parent of mout_i2s, but we reparent them to fout_epll and sclk_audio0,
respectively, when setting up the audio clock hierarchy.  They are
both then reparented on resume when the AUDSS registers are restored.
This, however, causes the machine to hang during resume.  It's not
clear what the exact cause is, but it's possible that they cannot be
reparented at the same time or that EPLL is still unstable.  Temporarily
reparenting mout_i2s to mout_audss across suspend/resume so that it is
not reparented when the AUDSS registers are restored appears to make
suspend/resume much more stable on Pit and Snow.

BUG=chrome-os-partner:20102
TEST=Can suspend/resume on Pit and Snow.  Audio still works.

Change-Id: I75e30b66bfc103de8aa83cc62b21a9f4a938cbb5
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58212
Reviewed-by: Simon Glass <sjg@chromium.org>